+++ config/dep_all.rs
@@ -2,11 +2,11 @@
 // Used from: ../build/features.rs::DEPENDENCY.features & ../utils/check.rs::DEP_ALL.
 //
 // NOTE: make sure to recompile `../utils/check.rs` after modifying this file.
-// 51 optional dependencies:
+// 48 optional dependencies:
 [
     "dep_allocator_api2", "dep_atomic", "dep_bumpalo", "dep_bytemuck",
     "dep_const_str", "dep_crossterm", "dep_fltk", "dep_flume", "dep_fontdue",
-    "dep_gilrs", "dep_hashbrown", "dep_image", "dep_itertools", "dep_jiff", "dep_js_sys",
+    "dep_gilrs", "dep_hashbrown", "dep_image", "dep_itertools", "dep_jiff",
     "dep_kira", "dep_libm", "dep_log", "dep_memchr", "dep_midir", "dep_miniquad",
 	"dep_nc", "dep_orion", "dep_portable_atomic", "dep_portable_atomic_util", "dep_pyo3",
     "dep_rand_core", "dep_rayon", "dep_raw_cpuid", "dep_regex_lite",
@@ -14,5 +14,5 @@
     "dep_serde", "dep_simdutf8", "dep_stringzilla", "dep_sysinfo", "dep_symphonia",
     "dep_tinyaudio", "dep_toml_edit", "dep_tokio", "dep_tracing",
     "dep_unicode_segmentation", "dep_unicode_width", "dep_ureq",
-    "dep_wasm_bindgen", "dep_web_sys", "dep_wide", "dep_winnow",
+    "dep_wide", "dep_winnow",
 ]